chapter-10/src/main/java/com/git/onedayrex/nettyaction/chapter10/coder/AbsIntegerEncoder.java
package com.git.onedayrex.nettyaction.chapter10.coder;

import io.netty.buffer.ByteBuf;
import io.netty.channel.ChannelHandlerContext;
import io.netty.handler.codec.MessageToMessageEncoder;

import java.util.List;

public class AbsIntegerEncoder extends MessageToMessageEncoder<ByteBuf> {

    @Override
    protected void encode(ChannelHandlerContext ctx, ByteBuf msg, List<Object> out) throws Exception {
        while (msg.readableBytes() >= 1) {
            int i = msg.readByte();
            int abs = Math.abs(i);
            out.add(abs);
        }
    }
}
